Resolve problems with dependencies
