+++ backend/autonext_server/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "maven": "^5.0.0"
+  }
+}
