+++ backend/autonext_server/src/main/java/com/autonext/code/autonext_server/configs/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.autonext.code.autonext_server.configs;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaRepositories
+public class JpaConfig {
+}
