+++ backend/autonext_server/src/main/java/com/autonext/code/autonext_server/services/AuthService.java
@@ -1,0 +1,21 @@
+// package com.autonext.code.autonext_server.services;
+
+// import org.springframework.security.authentication.AuthenticationManager;
+// import org.springframework.security.crypto.password.PasswordEncoder;
+// import org.springframework.stereotype.Service;
+
+// @Service
+// public class AuthService {
+
+//   private final UserRepository userRepository;
+//   private final PasswordEncoder passwordEncoder;
+//   private final JwtTokenProvider jwtTokenProvider;
+
+//   public AuthService(UserRepository userRepository, JwtService jwtService,
+//       AuthenticationManager authenticationManager, PasswordEncoder passwordEncoder) {
+//     this.userRepository = userRepository;
+//     this.jwtService = jwtService;
+//     this.authenticationManager = authenticationManager;
+//     this.passwordEncoder = passwordEncoder;
+//   }
+// }
