+++ backend/autonext_server/src/main/resources/application.properties
@@ -1,1 +1,15 @@
 spring.application.name=autonext_server
+spring.datasource.url=jdbc:mysql://localhost:3306/autoNextDB
+spring.datasource.username=backuser
+spring.datasource.password=backcontrasena
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+
+# Configuración de Hibernate
+spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.entity_manager_factory_interface=jakarta.persistence.EntityManagerFactory
+spring.jpa.show-sql=true
+
+# Configuración de logs
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
